+++ lib/screens/statistics/inventory_report_screen.dart
@@ -229,7 +229,7 @@ class InventoryReportScreen extends StatelessWidget {
       final headers = _buildColumns(selectedInventories).map((column) => column.label.toString()).toList();
       final csvData = [headers, ...reportData];
       final csv = const ListToCsvConverter().convert(csvData);
-      final directory = await getExternalStorageDirectory();
+      final directory = await getTemporaryDirectory();
       final path = '${directory?.path}/inventory_report_${DateTime.now().toIso8601String()}.csv';
       final file = File(path);
       await file.writeAsString(csv);